Rename frontend resource_management folder to resource-management
