+++ frontend/packages/app/src/app/pages/resource-management/team/components/expandView.tsx
