+++ frontend/packages/app/src/app/pages/resource-management/utils/group.ts
@@ -4,7 +4,7 @@
 import type {
   ResourceAllocationObjectProps,
   ResourceAllocationProps,
-} from "@/types/resource_management";
+} from "@/types/resource-management";
 import type {
   DateProps,
   EmployeeAllocationForDateProps,
