+++ frontend/packages/app/src/app/pages/resource-management/utils/helper.ts
