+++ frontend/packages/app/src/route.tsx
@@ -22,9 +22,9 @@ import { setViews } from "./store/view";
 const TimesheetComponent = lazy(() => import("@/app/pages/timesheet"));
 const HomeComponent = lazy(() => import("@/app/pages/home"));
 const TeamComponent = lazy(() => import("@/app/pages/team"));
-const ResourceTeamComponent = lazy(() => import("@/app/pages/resource_management/team"));
-const ResourceProjectComponent = lazy(() => import("@/app/pages/resource_management/project"));
-const ResourceTimeLineComponet = lazy(() => import("@/app/pages/resource_management/timeline"));
+const ResourceTeamComponent = lazy(() => import("@/app/pages/resource-management/team"));
+const ResourceProjectComponent = lazy(() => import("@/app/pages/resource-management/project"));
+const ResourceTimeLineComponet = lazy(() => import("@/app/pages/resource-management/timeline"));
 const EmployeeDetailComponent = lazy(() => import("@/app/pages/team/employee-detail"));
 const TaskComponent = lazy(() => import("@/app/pages/task"));
 const ProjectComponent = lazy(() => import("@/app/pages/project"));
